Refactor INDEX_COLUMN import path after constants relocation
Moved constants.js from InputEditor/components to input-editor root and updated all INDEX_COLUMN import paths accordingly. This improves code organization and simplifies imports across the input editor and map components.
